Iteration 01 complete
*Fixed version print w/o link
*All of custom functions and test inputs moved to separated file
*Unit tests added
*Json description added
*Tested on py 3.9.7..passed
*Iteration 01 complete
